First working version of the sketch
This sets up the display in the first boot, then goes to sleep. Every 5
seconds the display wakes up and adds a square in a random place
on the screen.

When testing `.hibernate()` between sleeps, noise appeared on the screen
before each partial update. Even after removing the call to `hibernate`
the problem persisted and only went away after reprogramming the
microcontroller many times. No idea of what happened there.

The standby current doesn't seem to go bellow 10mA with the ESP32 devkit
I'm using, apparently because most of the components there (such as the
USB->UART adapter) have a high quiescent current.