Game of life
This almost works. It's using a boolean array for the board, which
really limits the size of it (since it has to fit into RTC memory).
Also, for some reason the graphics don't show up as black, but actually
as noise, and they linger from one frame to the next, which means the
partial updates are not working as I expect them to.