Avoid noise that shows on partial update after sleep
The noise stays in the places where we used to have black and now have
white. This seems to be because of the difference from what was set in
the previous iteration and this one (the contents of the "old" memory in
the controler are not set).

By displaying the same image again before updating the board, we make
sure the "old" register has the correct information, and the partial
update works correctly.

There is still some flickering before the first display (of the old
board). This could probably be avoided by writing the buffer directly to
the "old" register in the controller. This would probably also allow us
to only do one partial update (the idea being: boot, write "old"
register, update board, do partial update).